requirements.txt
pandas==1.1.5
requests==2.18.4
altair_saver==0.5.0
iterfzf==0.5.0.20.0
loguru==0.5.3
altair==4.1.0
boardgamegeek2==1.0.1
tqdm==4.35.0
beautifulsoup4==4.9.3
fire
